BJS2-35145
refactoring code
